Update forgetpassword_server.php
I use session var to store email and hash in case there are password validation errors and user needs to resubmit the passwords. The change is to unset session vars in caes the user submits reset passwrods requests multiple times in a short period of time (each time a new hash will be generated).